Created the connection to the mongo db database.
For the datbase to wotk, we will need to use a .env file in which it will need to have the database URI.

The .env code needs to be called "MONGO_URI" for the connection to work.

Also added the dotenv and nodemon packges.

in packge.json added two scripts that we can use.

npm run test will execute the nodemon packege for testing purpouses.

npm run start will execute the code(Will be used for the render.com website).